Added withCredentials: true
When using the bootstrap.js alone (without Lineman) on top of our services.js you will end up with token missmatch exceptions all the time. Not too sure if this is linked to your ngix comment but regardless I found a patch to it. When using the $http you need to make sure to use the withCredentials